backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
langchain==0.3.19
langchain-google-genai==2.0.9
python-dotenv==1.0.1
pdfplumber==0.11.4
python-docx==1.1.2
pymupdf==1.25.3
